requirement.txt
stanfordcorenlp
librosa
epitran
tqdm
regex
transformers
bpemb
python_speech_features
torchsummary
torchaudio
